Both `exchange` and `purpose` need a special-purpose equality predicate, because the given equations and cards are actually sets, not ordered lists. Depending on how a player implementation de-serializes these pieces of data, the results differ on the surface but not actually semantically.
Equality on two purchase structs of the shape
```
(purchase cards points leftover-pebbles)
```
is given iff
  1. the two bunches of cards are equal as bags
     (the same card could show up twice)
  2. the points are the same
  3. the two bunches of leftover-pebbles are equal as bags

Equality on two exchange structs of the shape
```
(exchange oriented-equations some-purchase)
```
is given iff
  1. the two oriented-equations are equal as bags
     (the same oriented-equation could show up twice)
  2. the two purchases are equal as defined above.

Given these two notions of equality, the correct to "optimize"
(maximize points or number of cards) is to collect _sets_ of
equally optimal exchanges and purchases.

This means that different player implementations may submit different
orders of trades and different orders of card purchases.

THEOREM This will not affect the execution of the game.

PROOF

While the order of the trades may differ, the two make up the same set
of trades.

.: The two player implementations have the same bag of pebbles when
they begin to explore the purchase of cards.

While the order of card purchases may differ, the two implementations
end up buying the same cards and get the same number of points.

.: The two players have
- the same points,
- the same bag of pebbles
- the same number of points
when they are done with their turn.

Since points and pebbles are the only knowledge a referee has about a
player and player responses are judged based on this knowledge, the
future evolution of the game cannot change depending on the
implementation of the player's de-serialization of bag/set-like
entities. []